+++ ic2-reactors/src/main/groovy/com/gmail/dario/reactors/components/coolantcells/CoolantCell10K.groovy
@@ -5,7 +5,7 @@ import com.gmail.dario.reactors.components.ReactorComponent
 
 class CoolantCell10K extends HeatingObject {
 
-    final BigDecimal maxHeat = 10_000
+    final int maxHeat = 10_000
     
     @Override
     void tick() {/* do nothing*/}
