+++ ic2-reactors/src/main/groovy/com/gmail/dario/reactors/components/coolantcells/CoolantCell30K.groovy
@@ -5,7 +5,7 @@ import com.gmail.dario.reactors.components.ReactorComponent
 
 class CoolantCell30K extends HeatingObject {
 
-    final BigDecimal maxHeat = 30_000
+    final int maxHeat = 30_000
     
     @Override
     void tick() {/* do nothing*/}
