+++ ic2-reactors/src/main/groovy/com/gmail/dario/reactors/components/coolantcells/CoolantCell60K.groovy
@@ -5,7 +5,7 @@ import com.gmail.dario.reactors.components.ReactorComponent
 
 class CoolantCell60K extends HeatingObject {
 
-    final BigDecimal maxHeat = 60_000
+    final int maxHeat = 60_000
     
     @Override
     void tick() {/* do nothing*/}
