+++ ic2-reactors/src/main/groovy/com/gmail/dario/reactors/components/coolantcells/LZHCondensator.groovy
@@ -5,7 +5,7 @@ import com.gmail.dario.reactors.components.ReactorComponent
 
 class LZHCondensator extends HeatingObject {
 
-    final BigDecimal maxHeat = 100_000
+    final int maxHeat = 100_000
     
     @Override
     void tick() {/* do nothing*/}
