+++ ic2-reactors/src/main/groovy/com/gmail/dario/reactors/components/coolantcells/RSHCondensator.groovy
@@ -5,7 +5,7 @@ import com.gmail.dario.reactors.components.ReactorComponent
 
 class RSHCondensator extends HeatingObject {
 
-    final BigDecimal maxHeat = 20_000
+    final int maxHeat = 20_000
     
     @Override
     void tick() {/* do nothing*/}
