+++ ic2-reactors/src/main/groovy/com/gmail/dario/reactors/components/fuelcells/QuadUraniumCell.groovy
@@ -2,7 +2,7 @@ package com.gmail.dario.reactors.components.fuelcells
 
 class QuadUraniumCell extends UraniumCell {
 
-    final BigDecimal numberOfRods = 4
+    final int numberOfRods = 4
     final BigDecimal euGenerated = 60
 
     QuadUraniumCell() {
