+++ ic2-reactors/src/main/groovy/com/gmail/dario/reactors/components/fuelcells/SingleUraniumCell.groovy
@@ -2,7 +2,7 @@ package com.gmail.dario.reactors.components.fuelcells
 
 class SingleUraniumCell extends UraniumCell {
 
-    final BigDecimal numberOfRods = 1
+    final int numberOfRods = 1
     final BigDecimal euGenerated = 5
 
     SingleUraniumCell() {
