+++ ic2-reactors/src/main/groovy/com/gmail/dario/reactors/components/heatexchangers/AdvancedHeatExchanger.groovy
@@ -2,11 +2,11 @@ package com.gmail.dario.reactors.components.heatexchangers
 
 class AdvancedHeatExchanger extends HeatMover {
 
-    final BigDecimal maxHeat = 10_000
+    final int maxHeat = 10_000
 
     @Override
     void tick() {
-        transferHeatWithConnectedComponents(24.0)
-        transferHeatWithReactor(8.0)
+        transferHeatWithConnectedComponents(24)
+        transferHeatWithReactor(8)
     }
 }
