+++ ic2-reactors/src/main/groovy/com/gmail/dario/reactors/components/heatexchangers/ComponentHeatExchanger.groovy
@@ -2,10 +2,10 @@ package com.gmail.dario.reactors.components.heatexchangers
 
 class ComponentHeatExchanger extends HeatMover {
 
-    final BigDecimal maxHeat = 5_000
+    final int maxHeat = 5_000
 
     @Override
     void tick() {
-        transferHeatWithConnectedComponents(36.0)
+        transferHeatWithConnectedComponents(36)
     }
 }
