+++ ic2-reactors/src/main/groovy/com/gmail/dario/reactors/components/heatexchangers/HeatExchanger.groovy
@@ -2,12 +2,12 @@ package com.gmail.dario.reactors.components.heatexchangers
 
 class HeatExchanger extends HeatMover {
 
-    final BigDecimal maxHeat = 2_500
+    final int maxHeat = 2_500
 
     @Override
     void tick() {
-        transferHeatWithConnectedComponents(12.0)
-        transferHeatWithReactor(4.0)
+        transferHeatWithConnectedComponents(12)
+        transferHeatWithReactor(4)
     }
     
 }
