+++ ic2-reactors/src/main/groovy/com/gmail/dario/reactors/components/heatexchangers/ReactorHeatExchanger.groovy
@@ -2,10 +2,10 @@ package com.gmail.dario.reactors.components.heatexchangers
 
 class ReactorHeatExchanger extends HeatMover {
 
-    final BigDecimal maxHeat = 5_000
+    final int maxHeat = 5_000
 
     @Override
     void tick() {
-        transferHeatWithReactor(72.0)
+        transferHeatWithReactor(72)
     }
 }
